* commit '144e23cf4293c9d83989b2668bc670b99ea85950': (341 commits)
  Stop lowering external id when searching
  fix binds and SQL-statement
  fix binds for mac_adr
  Convert spread to constant when filtering groups
  add a t to avoid confusion
  Fix broken SQL statement
  Fix renaming by name
  Fix malformed sql query
  Fix broken SQL-query
  Remove member names from email
  Rewrite two similar SQL-statements in explicit terms
  reverting a malformed pep8-improvement
  Add a safety measure for the future
  Fix typo
  Remove reference to removed variable
  Fix mistake
  Redefine strings for SQL statement
  Fix a typo
  Fix a bug due to NotSet
  Add more information and info link
  ...

+++ Cerebrum/modules/dns/AAAARecord.py
@@ -2,6 +2,7 @@
 from Cerebrum.Entity import Entity
 from Cerebrum.modules.dns.DnsOwner import DnsOwner
 
+
 class AAAARecord(Entity):
     """``AAAARecord(Entity)`` is used to store information
     about AAAA-records in the dns_a_record table.  It uses the standard
@@ -27,7 +28,7 @@ class AAAARecord(Entity):
             Entity.populate(self, self.const.entity_dns_aaaa_record)
         try:
             if not self.__in_db:
-                raise RuntimeError, "populate() called multiple times."
+                raise RuntimeError("populate() called multiple times.")
         except AttributeError:
             self.__in_db = False
         for k in locals().keys():
@@ -36,10 +37,11 @@ class AAAARecord(Entity):
 
     def __eq__(self, other):
         assert isinstance(other, AAAARecord)
-        if (self.dns_owner_id != other.dns_owner_id or
-            self.ipv6_number_id != other.ipv6_number_id or
-            self.ttl != other.ttl or
-            self.mac != other.mac):
+        if (
+                self.dns_owner_id != other.dns_owner_id or
+                self.ipv6_number_id != other.ipv6_number_id or
+                self.ttl != other.ttl or
+                self.mac != other.mac):
             return False
         return True
 
@@ -48,42 +50,57 @@ class AAAARecord(Entity):
         if not self.__updated:
             return
         is_new = not self.__in_db
-        cols = [('aaaa_record_id', ':e_id'),
-                ('entity_type', ':e_type'),
-                ('dns_owner_id', ':dns_owner_id'),
-                ('ipv6_number_id', ':ipv6_number_id'),
-                ('ttl', ':ttl'),
-                ('mac', ':mac')]
-        binds = {'e_id': self.entity_id,
-                 'e_type' : int(self.const.entity_dns_aaaa_record),
+        binds = {'aaaa_record_id': self.entity_id,
+                 'entity_type': int(self.const.entity_dns_aaaa_record),
                  'dns_owner_id': self.dns_owner_id,
                  'ipv6_number_id': self.ipv6_number_id,
-                 'ttl' : self.ttl,
+                 'ttl': self.ttl,
                  'mac': self.mac}
-
+        defs = {'tc': ', '.join(x for x in sorted(binds)),
+                'tb': ', '.join(':{0}'.format(x) for x in sorted(binds)),
+                'ts': ', '.join('{0}=:{0}'.format(x) for x in binds
+                                if x != 'aaaa_record_id')}
         if is_new:
-            self.execute("""
-            INSERT INTO [:table schema=cerebrum name=dns_aaaa_record] (%(tcols)s)
-            VALUES (%(binds)s)""" % {'tcols': ", ".join([x[0] for x in cols]),
-                                     'binds': ", ".join([x[1] for x in cols])},
-                         binds)
-            self._db.log_change(self.dns_owner_id, self.clconst.aaaa_record_add,
+            insert_stmt = """
+            INSERT INTO [:table schema=cerebrum name=dns_aaaa_record] (%(tc)s)
+            VALUES (%(tb)s)""" % defs
+            self.execute(insert_stmt, binds)
+            self._db.log_change(self.dns_owner_id,
+                                self.clconst.aaaa_record_add,
                                 self.ipv6_number_id)
         else:
-            self.execute("""
-            UPDATE [:table schema=cerebrum name=dns_aaaa_record]
-            SET %(defs)s
-            WHERE aaaa_record_id=:e_id""" % {'defs': ", ".join(
-                ["%s=%s" % x for x in cols])}, binds)
-            self._db.log_change(self.dns_owner_id,
-                    self.clconst.aaaa_record_update, self.ipv6_number_id)
+            exists_stmt = """
+              SELECT EXISTS (
+                SELECT 1
+                FROM [:table schema=cerebrum name=dns_aaaa_record]
+                WHERE aaaa_record_id=:aaaa_record_id AND
+                      entity_type=:entity_type AND
+                     (ttl is NULL AND :ttl is NULL OR ttl=:ttl) AND
+                     (dns_owner_id is NULL AND :dns_owner_id is NULL OR
+                       dns_owner_id=:dns_owner_id) AND
+                     (ipv6_number_id is NULL AND :ipv6_number_id is NULL OR
+                       ipv6_number_id=:ipv6_number_id) AND
+                     (mac is NULL AND :mac is NULL OR
+                       mac=:mac)
+              )
+            """
+            if not self.query_1(exists_stmt, binds):
+                # True positive
+                update_stmt = """
+                UPDATE [:table schema=cerebrum name=dns_aaaa_record]
+                SET %(ts)s
+                WHERE aaaa_record_id=:aaaa_record_id""" % defs
+                self.execute(update_stmt, binds)
+                self._db.log_change(self.dns_owner_id,
+                                    self.clconst.aaaa_record_update,
+                                    self.ipv6_number_id)
         del self.__in_db
-        
         self.__in_db = True
         self.__updated = []
         return is_new
 
     def find_by_name(self, name):
+        """Find DNS by owner's name"""
         # Will result in a number of queries, but it is not expected
         # that find_by_name will be used in performance-intesive
         # queries.
@@ -92,6 +109,7 @@ class AAAARecord(Entity):
         self.find_by_dns_owner_id(dns.entity_id)
 
     def find_by_dns_owner_id(self, dns_owner_id):
+        """Find DNS by owner_id"""
         # May throw TooManyRows error, which callee should handle
         aaaa_record_id = self.query_1("""
         SELECT aaaa_record_id
@@ -100,30 +118,34 @@ class AAAARecord(Entity):
         self.find(aaaa_record_id)
 
     def find_by_ip(self, ipv6_number_id):
+        """Find DNS by IPv6"""
         # May throw TooManyRows error, which callee should handle
         aaaa_record_id = self.query_1("""
         SELECT aaaa_record_id
         FROM [:table schema=cerebrum name=dns_aaaa_record]
-        WHERE ipv6_number_id=:ipv6_number_id""", {'ipv6_number_id': ipv6_number_id})
+        WHERE ipv6_number_id=:ipv6_number_id""",
+                                      {'ipv6_number_id': ipv6_number_id})
         self.find(aaaa_record_id)
 
     def find_by_owner_and_ip(self, ipv6_number_id, dns_owner_id):
+        """Find DNS by owner_id and IPv6"""
         aaaa_record_id = self.query_1("""
         SELECT aaaa_record_id
         FROM [:table schema=cerebrum name=dns_aaaa_record]
-        WHERE dns_owner_id=:dns_owner_id AND ipv6_number_id=:ipv6_number_id""", {
-            'dns_owner_id': dns_owner_id,
-            'ipv6_number_id': ipv6_number_id})
+        WHERE dns_owner_id=:dns_owner_id AND ipv6_number_id=:ipv6_number_id""",
+                                      {'dns_owner_id': dns_owner_id,
+                                       'ipv6_number_id': ipv6_number_id})
         self.find(aaaa_record_id)
 
     def find(self, aaaa_record_id):
         self.__super.find(aaaa_record_id)
-
-        (self.ipv6_number_id, self.ttl, self.mac, self.dns_owner_id
-         ) =  self.query_1("""
+        (self.ipv6_number_id,
+         self.ttl, self.mac,
+         self.dns_owner_id) = self.query_1("""
         SELECT ipv6_number_id, ttl, mac, dns_owner_id
         FROM [:table schema=cerebrum name=dns_aaaa_record]
-        WHERE aaaa_record_id=:aaaa_record_id""", {'aaaa_record_id' : aaaa_record_id})
+        WHERE aaaa_record_id=:aaaa_record_id""",
+                                           {'aaaa_record_id': aaaa_record_id})
         dns = DnsOwner(self._db)
         dns.find(self.dns_owner_id)
         self.name = dns.name
@@ -138,11 +160,23 @@ class AAAARecord(Entity):
         """Deletion of aaaa-records should be done through the
         IntegrityHelper to avoid leaving entries in ipv6_number that has
         no FKs to it"""
-        self.execute("""
-        DELETE FROM [:table schema=cerebrum name=dns_aaaa_record]
-        WHERE aaaa_record_id=:e_id""", {'e_id': self.entity_id})
-        self._db.log_change(self.dns_owner_id, self.clconst.aaaa_record_del,
-                            self.ipv6_number_id)
+        binds = {'aaaa_record_id': self.entity_id}
+        exists_stmt = """
+          SELECT EXISTS (
+          SELECT 1
+          FROM [:table schema=cerebrum name=dns_aaaa_record]
+          WHERE aaaa_record_id=:aaaa_record_id
+        )
+        """
+        if self.query_1(exists_stmt, binds):
+            # True positive
+            delete_stmt = """
+            DELETE FROM [:table schema=cerebrum name=dns_aaaa_record]
+            WHERE aaaa_record_id=:aaaa_record_id"""
+            self.execute(delete_stmt, binds)
+            self._db.log_change(self.dns_owner_id,
+                                self.clconst.aaaa_record_del,
+                                self.ipv6_number_id)
         self.__super.delete()
 
     def list_ext(self, ip_number_id=None, dns_owner_id=None, start=None,
@@ -166,9 +200,8 @@ class AAAARecord(Entity):
              [:table schema=cerebrum name=dns_owner] d,
              [:table schema=cerebrum name=entity_name] en
         WHERE %s
-        ORDER BY aaaa_record_id""" % where, {
-            'ipv6_number_id': ip_number_id,
-            'dns_owner_id': dns_owner_id,
-            'start': start, 'stop': stop,
-            'zone': zone} )
-
+        ORDER BY aaaa_record_id""" % where,
+                          {'ipv6_number_id': ip_number_id,
+                           'dns_owner_id': dns_owner_id,
+                           'start': start, 'stop': stop,
+                           'zone': zone})
